анбоксинг

+++ src/main/java/got/client/GOTReflectionClient.java
@@ -60,7 +60,7 @@ public class GOTReflectionClient {
 	public static float getFOVModifier(EntityRenderer renderer, float tick, boolean flag) {
 		try {
 			Method method = GOTReflection.getPrivateMethod(EntityRenderer.class, renderer, new Class[]{Float.TYPE, Boolean.TYPE}, "getFOVModifier", "func_78481_a");
-			return (Float) method.invoke(renderer, Float.valueOf(tick), flag);
+			return (Float) method.invoke(renderer, tick, flag);
 		} catch (Exception e) {
 			GOTReflection.logFailure(e);
 			return 0.0f;
