+++ src/main/java/got/common/entity/essos/legendary/warrior/GOTEntityAsshaiArchmag.java
@@ -57,7 +57,7 @@ public class GOTEntityAsshaiArchmag extends GOTEntityHumanBase {
 	@Override
 	public void entityInit() {
 		super.entityInit();
-		dataWatcher.addObject(17, Byte.valueOf((byte) 0));
+		dataWatcher.addObject(17, (byte) 0);
 	}
 
 	@Override
